feat(clients): Config files
Configuration files added for database connection
in different environments.

Like:

if you are using docker-compose for the environment of the service,
the service must use the `config.docker.json`, that's because in the docker-compose.yml is
defined the DOCKER_ENV environment variable.

if there is not set a DOCKER_ENV env var, the service will use the `config.json` for
database connection.

but if is set the environment vars DB_HOST, and DB_PORT, all this is disconsidered
and then it will use the environment vars.